Update the SED used to edit shairport-sync.conf
This update fixes an issue where in the 4.1.1 conf there are two occurrences of audio_backend_buffer_desired_length_in_seconds param where the second occurrence is in a comment line.

This causes the original SED to also edit the second occurrence so it too becomes an active param but with a trailing single quote.

The trailing quote causes shairport-sync to abort with a "syntax error in conf file" message.